refactor(hfresh): fold bucketRef and the teardown tests into the store files
bucketRef is the primitive the posting store is built on and had a file to
itself; move it to the top of store.go, ahead of its first user. Its tests and
the rest of the teardown coverage move into store_test.go, and the two files
that held them are gone.

Code and tests are unchanged; the benchmark moves below the tests so the file
reads tests-then-benchmark.
